+++ app/src/main/java/com/london/tudee/presentation/components/navigation_bar_bottom/BottomNavDefaults.kt
@@ -1,4 +1,4 @@
-package com.london.tudee.presentation.components.bottom_navigation_bar
+package com.london.tudee.presentation.components.navigation_bar_bottom
 
 import com.london.tudee.R
 import com.london.tudee.presentation.navigation.Screen
