+++ app/src/main/java/com/london/tudee/presentation/screens/task/TasksScreen.kt
@@ -44,13 +44,13 @@ import androidx.compose.ui.unit.dp
 import com.london.tudee.R
 import com.london.tudee.domain.entities.Category
 import com.london.tudee.domain.entities.Task
-import com.london.tudee.presentation.components.EmptyTasksScreen
 import com.london.tudee.presentation.components.SnackBar
 import com.london.tudee.presentation.components.buttons.TudeeFloatingActionButton
 import com.london.tudee.presentation.components.date.DateItem
 import com.london.tudee.presentation.components.date.TudeeDatePicker
 import com.london.tudee.presentation.components.tabs.TabItem
 import com.london.tudee.presentation.components.tabs.TudeeTabLayoutWithPager
+import com.london.tudee.presentation.components.task.EmptyTasksScreen
 import com.london.tudee.presentation.components.task.SwipeToDeleteTask
 import com.london.tudee.presentation.design_system.theme.ThemePreviews
 import com.london.tudee.presentation.design_system.theme.TudeeTheme
@@ -182,7 +182,7 @@ fun TasksContent(
                     onClickDay = onDayClick,
                 )
             }
-        ) { page, tasks ->
+        ) { _, tasks ->
             if (tasks.isEmpty()) {
                 Box(
                     modifier = Modifier
@@ -216,8 +216,8 @@ fun TasksContent(
 
     if (showDatePicker)
         TudeeDatePicker(
-            onDateSelected = { date ->
-                date?.let { onDateSelected(it) }
+            onDateSelected = { selectedDate ->
+                selectedDate?.let { onDateSelected(it) }
                 showDatePicker = false
             },
             onDismiss = { showDatePicker = false }
